Testes integrados

+++ src/controllers/experiment/experiment.controller.ts
@@ -22,6 +22,8 @@ export class ExperimentController {
     let orderBy: object | any;
     parameters.AND = [];
     try {
+      console.log('options');
+      console.log(options);
       if (options.filterRepetitionFrom || options.filterRepetitionTo) {
         if (options.filterRepetitionFrom && options.filterRepetitionTo) {
           parameters.repetitionsNumber = JSON.parse(`{"gte": ${Number(options.filterRepetitionFrom)}, "lte": ${Number(options.filterRepetitionTo)} }`);
@@ -290,23 +292,26 @@ export class ExperimentController {
 
   async delete(data: any) {
     try {
+      const experimentGenotipeController = new ExperimentGenotipeController();
       const experimentExist: any = await this.getOne(Number(data.id));
       if (!experimentExist) return { status: 404, message: 'Experimento não encontrado' };
       if (experimentExist?.status === 'PARCIALMENTE ALOCADO' || experimentExist?.status === 'TOTALMENTE  ALOCADO') return { status: 400, message: 'Não é possível deletar.' };
+      const { status } = await experimentGenotipeController.deleteAll(data.id);
+      if (status === 200) {
+        const response = await this.experimentRepository.delete(Number(data.id));
+        const {
+          response: assayList,
+        } = await this.assayListController.getOne(Number(experimentExist?.idAssayList));
+        if (!assayList?.experiment.length) {
+          await this.assayListController.update({
+            id: experimentExist?.idAssayList,
+            status: 'IMPORTADO',
+          });
+        }
 
-      const response = await this.experimentRepository.delete(Number(data.id));
-      const {
-        response: assayList,
-      } = await this.assayListController.getOne(Number(experimentExist?.idAssayList));
-      if (!assayList?.experiment.length) {
-        await this.assayListController.update({
-          id: experimentExist?.idAssayList,
-          status: 'IMPORTADO',
-        });
-      }
-
-      if (response) {
-        return { status: 200, message: 'Experimento excluído' };
+        if (response) {
+          return { status: 200, message: 'Experimento excluído' };
+        }
       }
       return { status: 404, message: 'Experimento não excluído' };
     } catch (error: any) {
